requirements.txt
Flask==3.0.3
pillow==11.0.0
requests==2.32.3